Added look command
I added a look command to the idle loop.
I'm thinking I need to start over and refactor everything. The current
structure is getting unweildy for the changes I want to make. All the
props are going to need a 'container' field so that it knows what it is
in.